+++ SHMR_Finance_App/SHMR_Finance_App/Common/SpoilerAnimation/SpoilerModifier.swift
