+++ SHMR_Finance_App/SHMR_Finance_App/Common/Views/View+ErrorAlert.swift
