+++ SHMR_Finance_App/SHMR_Finance_App/Models/BankAccount.swift
@@ -23,32 +23,4 @@ enum Currency: String {
     case EUR = "€"
 }
 
-/*
- enum Currency: String {
- case rub = "RUB"
- case usd = "USD"
- case eur = "EUR"
- 
- var symbol: String {
- switch self {
- case .rub:
- return "₽"
- case .usd:
- return "$"
- case .eur:
- return "€"
- }
- }
- 
- var fullName: String {
- switch self {
- case .rub:
- return "Российский рубль ₽"
- case .usd:
- return "Американский доллар $"
- case .eur:
- return "Евро €"
- }
- }
- }
- */
+
